Initial (apparently working) EFI stub in go, with some comments.
This is based on tamago - may try to get it with official go, but no
rush or need, tamago seems good enough.

While original C works pretty well - and the Zig replacement is cleaner
and easier to build - I still like go more and it is easier to
experiment. May port it back to zig - but I think the C code will just
remain for reference